Main Changes: 1)Implemented the available/locked on both the characters class and objects class. This was relatively easy, as it just compares the special points to an x value (depends on each character/object), if its larger then the set value, It displays 'Available', else 'Locked'. 2)Fixed the issue with the 'stickman theme object' not being displayed when selected. This issue was down to the missing outlet connection. 3)Implemented a new userdefault for the final value that is used to calculate the average score. The issue I had before, was that I was send continuous Svalue to allValues. As overtime the score changed the allValues changed to allValues+Svalues. Which is not what I wanted. Since I needed the final Svalue and not the Svalue that changed every time. Hence this was easily solved with finalValue. 4)I was also had issues with displaying the average value in statistics class. The main problem was the definition of the type of verbal. Since I was trying to display a long variable, whilst calling a float value.

+++ Game/GameOver.h
@@ -13,6 +13,6 @@
 @property (strong, nonatomic) IBOutlet UILabel *CurrentScore;
 @property (strong, nonatomic) IBOutlet UILabel *BestScore;
 @property float AverageScore;
-@property int AllScores;
+@property float AllScores;
 
 @end
